requirements.txt
numpy==1.21.4
paddlepaddle==2.2.0
Cython==0.29.24
